Add files via upload
Currently has working movement, termination on hitting borders, changing of the snake's tail based on direction, etc. Currently working on creating a 2D array to store previously visited locations. This means changing the while(1) loop to an infinite for loop ie: for(int i = 0; i >= 0; i++) to keep track of which move you're on. This 2D array would have its height be the number of moves you're on and its width be 1 (index 0 is row and index one is col). Then another check would be added to see if the next desired move has an x,y coordinate already matching one of the elements in the visited squares array, and if it does, run the kill_snake() method and game over.