Add files via upload
This version has working collision detection for when the snake hits the borders, but also the rest of the snake's tail. Next on the list is managing the size of the snake, the length of its tail, having it actually move, rather than just grow forever, and inserting the trophy mechanics.